Added axis-aligned bounding boxes Refactored Octree and SkinnedModel::AutoSkinModel to use AABBs Added defines to get rid of min/max macros, and to make math header include math constants Fix warnings because M_PI is officially a double

+++ DestructibleTerrain/DTScreen.cpp
@@ -73,7 +73,7 @@ namespace DestructibleTerrain
 			}
 
 			// if there are somehow no materials in or around this node, don't bother with these steps
-			if(unsigned int count = min(4, total_materials.size()))
+			if(unsigned int count = min(4u, (unsigned int)total_materials.size()))
 			{
 				unsigned char mat_types[4];
 				unsigned int mat_weights[4];
@@ -261,7 +261,7 @@ namespace DestructibleTerrain
 			glClearColor(0.0f, 0.0f, 0.0f, 1.0f);
 			glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
 
-			camera_ori = Quaternion::FromPYR(0, -yaw, 0) * Quaternion::FromPYR(pitch, 0, 0) * Quaternion::FromPYR(0, M_PI, 0);
+			camera_ori = Quaternion::FromPYR(0, -yaw, 0) * Quaternion::FromPYR(pitch, 0, 0) * Quaternion::FromPYR(0, float(M_PI), 0);
 			
 			Mat3 rm = camera_ori.ToMat3();
 			Vec3 forward(Vec3::Normalize(rm * Vec3(0, 0, -1)));
